Document HttpLayer
References #254
